+++ kafka-connect-s3/src/main/java/io/confluent/connect/s3/S3SinkConnector.java
@@ -97,9 +97,13 @@ public class S3SinkConnector extends SinkConnector {
       for (ConfigValue configValue : config.configValues()) {
         if (configValue.name().equals(S3SinkConnectorConfig.S3_BUCKET_CONFIG)) {
           String bucketName = configValue.value().toString();
-          configValue.addErrorMessage(String.format("The S3 bucket named %s could not be found."
-                  + " Check the S3 credentials and verify the %s bucket exists.",
-              bucketName, bucketName));
+          configValue.addErrorMessage(
+              String.format("The S3 bucket named %s could not be found."
+                      + " Check the S3 credentials and verify the %s bucket exists.",
+                  bucketName,
+                  bucketName
+              )
+          );
         }
       }
     }
